requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.1.3
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
python-dotenv==0.20.0
requests==2.28.1
SQLAlchemy==1.4.39
urllib3==1.26.11
Werkzeug==2.2.0
zipp==3.8.1
